World say hello to JUMMP
This first code dump contains the code generated by
* grails create-app
* installing plugin spring-security-core
* installing plugin spring-security-ldap

Furhtermore it has some working code to bootstrap the initial
application configuration. All web requests of the first run
are redirected to a SetupController to configure the MySQL DB
and the authentication backend. JUMMP already supports
authenticating against database and LDAP backend. In the second
application startup all requests can be redirected to create an
admin account stored in the database. This admin is able to login.