Add springsecurity ui plugin.
This plugin allows to configure/view every aspect of springsecurity framework.
All controllers are in the plugin and are not protected by a secured annotation.
Because of that a static mapping is used in the configuration.
The plugin has quite some dependencies, one of them is jquery, so now we have
an ajax framework.