Adding a VcsManager interface.
The interface needs to be implemented by a manager providing access to a version control system (e.g. SVN).
It provides the required methods to add files to the repository and to retrieve files again. It is designed
with a checked out working copy in mind.

In order to have a working error handling some exception classes are also added. There is a JummpException
which is the base to all exceptions inside jummp and a base VcsException with some more specific inheriting
exceptions.