Move subversion backend into a plugin.
This adds the first ever jummp plugin. It's a normal grails plugin.
Be aware that this introduces a circular module dependency. The plugin
depends on interfaces and classes defined in the core and to run the core
the plugin is required. A proper build system will fix this problem.

Currently you will get compile errors when trying to build the plugin.