Using CMake as build system.
CMake allows to define proper build dependencies without the need for complex xml files as in ant.
Small modules can be defined in own files living in subdirectories. This makes modularization of the
build and inter-module dependencies much easier to define. There is not one huge build file, but a set
of build files for one specific task (e.g. building a plugin).
CMake uses out of source builds, so the source tree does not get polluted.

At the moment grails commands are still executed in the source tree, so this is still suboptimal and
CMake has minor issues when building Java sources.