Adding a git backend.
This is an alternative plugin to the subversion one. It also implements the VcsManager
and provides exactly the same functionality, but just with git. The idea behind
the plugin is to show that it is possible to have several plugins providing the same
functionality. In future it will be possible to select which plugin to use.

The git plugin uses the JGit library written for eclipse, the most current version
is included in the lib folder. The code is not yet tested for work with remote
repositories, but just having a git repository is tested via unittests.