Adding a VcsService resolving the concrete Plugin.
The VcsService resolves the vcs backend provided by a plugin.
Therefore a new interface is added which is implemented by a service in the plugin.
The interface provides a method to get a VcsManager.

The whole thing is not yet working - only committing to get a clean index.
Yeah git has advantages :-(