Move the core API into an own plugin.
This solves the problem of circular dependencies. The core application depended on the plugins and the plugins depended on the core.
The circular dependencies caused compile errors which were difficult to circumvent.
With this core-api plugin all plugins can depend on the api and the application itself as well. So both application and all plugins
can be compiled without dependency issues.