Adding a ModelService to access models and revisions in a secure way.
This service provides the methods to get all models, the revisions of a model,
the latest revision and so on. Not all methods are implemented yet, those which
are have a corresponding integration test.

The service uses Spring Security ACL to ensure who has access to a model or revision.
All other parts of JUMMP which need Models or Revisions should use this service and
not use GORM directly. GORM will circumvent the security completely (see examples in
the integration tests).