First part of test for addRevision.
The test so far show a problem with the current implementation:
when a revision can be added to the VCS but not to the database, the
VCS will come into an inconsistent state. It also seems difficult to test
this condition.

The implementation is currently missing a way to grant the read right
to all users who have read permission on the model.