Properly handle situation that config pluginServiceName is not set.
In case the ConfigObject does not contain a key, the value is a map.
But when it is set the value evaluates to whatever it is, in our case
to a String. Therefore we need to use a def and check if it is a String.

Thanks to Coverage test to highlight that the else branch has not been
executed.