Create and store a new model.
The method is currently written to only allow storing a new Model, if the user is logged-in.
To support the possible anonymous submission, not many changes are required. The security check
needs to be removed and instead of the current user becoming owner of the model the curators need
to become owner and the curation process needs to be invoked.