Adding JMS integration to Jummp.
JMS will in future be used for the communication between the core and the web application/ web services.
It is added as a new plugin which includes the library to access activemq and a small addition to the
Grails JmsPlugin to do synchronous API calls (which we need for direct communication).

Nevertheless the core application also needs to include the JmsPlugin for compilation issues.

To use JMS Apache ActiveMq in version 4.5.2 (current at the time of the inclusion) needs to be running.
Jummp connects to the server running on localhost. This (of course) needs to become configurable.