Adding a JmsAdapterService.
This service is an adapter to make all core functionality available through JMS.
Sending JMS methods comes with some obstacles:
 * messages only take one argument. Multiple Arguments need to be wrapped in a list
 * Synchronous methods time out if the called method does not return a value.
 * A timed out method returns always null
 * Exceptions are not passed to the callee, it ends in a timeout
 * Calling methods with no arguments is not possible, it has to be exactly one

The JmsAdapterService tries to hide these ugly parts. How it can be used can be seen in the
integration test. Each method validates the passed in message and will return (not thrown) an
IllegalArgumentsException if the message format is incorrect. In general the methods represent
the ModelService but take an additional first argument: the current Authentication.

To gain an Authentication a method is added to verify an Authentication. This can be used
to pass in a UsernamePasswordAuthenticationToken and if it validates, it can be passed in
subsequent calls. The core does not further validate the Authentication. It completely trusts
the callee.

Each method sets the passed in Authentication in the current SecurityContext, so that
Spring Security can verify the service calls against the Authentication. The methods will
clear the Authentication before returning. Unfortunately using any kind of Aspect to set
the Authentication did not work.

The delegation to the actual method of the ModelService is wrapped inside a try/catch/finally
block and each expected Exception is catched and returned. The idea is that a similar adapter
exists on the callee site, taking care of validating the returned value and rethrowing any
returned exception. Runtime Exceptions are not catched and will result in a returned null
value, therefore the callee has also to check whether the returned value is null and should
throw an Exception.

I must say that the integration of JMS is rather ugly. It works and the bloody details can be
hidden through the services. Nevertheless I will look into using a different IPC mechanism which
provides a "normal" API. Maybe D-Bus...