Workaround the problem that GrailsUser is not serializable.
Because GrailsUser is not serializable we cannot pass the Authentication to a different running instance through JMS.
The GrailsUser is wrapped into a SerializableGrailsUser extending the same base class as GrailsUser and implementing
the additions of GrailsUser.

After the Authentication is initially generated a new one is created and the custom serialized version of GrailsUser
replaces the principal. The same needs to be done for the AuthenticationException which has the GrailsUser as the
optional extraInformation.

Further methods do not need to be adjusted as the serialized version is passed in, so the thrown AuthenticationException
will have that one, too (at least I expect this).