Command objects to pass Model and Revision through JMS.
The general idea of this change is to decouple the returned data completely from the Database.
By using a command object it is clear, that it is not possible to just modify the objects and
call save on them which would cause nice Hibernate exceptions if done on a real DOM class.
Also this allows to keep the Model and Revision domain classes in the core and not in a plugin.
The command objects can hide the details which should not be visible at all. That is the VCS
information and all Revisions in the Model. To get the Revisions the services have to be used
which ensure security.

The Model and Revision provide a method toCommandObject to gain a matching object. In the
other direction the Model.get(commandObject.id) can be used. Though a static fromCommandObject
might be an idea.

This changes the signatures of the JmsAdapterService therefore the tests are adjusted.
The ModelState is also moved to the core-api plugin to be able to be referenced in the CommandObject.