CoreAdapterService for communication with Core via JMS.
This service wraps all the methods exported by the core through the JmsAdapterService into
a useful API. It takes care of setting the Authentication, re-throwing received Exceptions
and validating the returned value (if not matching JummpExceptions are thrown).

The integration test for this service needs to be operated against a running core application
in test-core mode. It is important to restart the core application after the test execution
has finished. Due to the fact that the persistancy is in a different application the database
and VCS is in deed changed in each test and not reset.