ModelFormat becomes a Domain Class.
Instead of an enum the ModelFormat is now a Domain class with instances stored in the database.
The default "UNKNOWN" format is created while bootstrapping the application and plugins are allowed
to add their own formats through the ModelFileFormatService. This service can do runtime decisions
on which plugin handles a format and not a compile time decision any more.

With this change the core application should not depend on any JUMMP plugin anymore.