Improved build to exclude plugins.
Using ant build to generate the war. Before it copies all required
sources to a build directory and there is a task to interactively
select the plugins to include. If a plugin is excluded its directory
is deleted by a gant script.

All grails plugins not required in the productive build (e.g codenarc)
are also removed before building.