Adding a web-application project.
This web-application is a standalone grails application communicating
with the standalone core application through JMS. It has an own
AuthenticationProvider which authenticates a UsernamePasswordAuthenticationToken
against the core application through JMS.

The Login process is already working, the user gets authenticated with the
correct username/password and an error message if authentication fails.

Now we can start adding some UI to the application.