Adding the last modification date to ModelTransportCommand.
Using the upload date of the latest Revision as the last modification date of the
Model. It is to be decided whether the ACL protected Revision should be used. A test
so far resulted in a crash when trying to access the ACL protected ModelService from
Model.

The need for this property in ModelTransportCommand is that BioModels DB displays
the last modification date in the overview of models, so we need it to be included
in the list returned by getAllModels.