The first controller to list models.
The controller provides access to all models in a secure way. The models
are fetched asynchronous and passed to the view in a JSON structure prepared
for jQuery DataTables plugin.

The view renders the models in a way similar to the existing model overview
in BioModels DB. All text elements are fetched from messages.properties to
allow customization. Not yet working is the sorting on table columns and
the JavaScript code should be moved into an own file.