Implement Jummp Logging Event Framework
Whenever a method invokation may be logged it can be annotated with the new
PostLogging annotation taking a parameter of the type of operation the method
performs. There is an advice which runs after-returning of the method and
generates an ApplicationEvent (LoggingEvent) with inoformation about the
user who called the method, the method itself its arguments and return
value. The event is broadcasted to the event handling system of spring.
Each interested party (e.g. a plugin) can implement an ApplicationListener
to receive the events and handle them.

The new LoggingEvent system should make it possible to intercept really all
methods which may change the state of the application and by that it should
be possible to log the events in an appropriate way. By using an aspect
together with an annotation it can be made clear in the code which areas
might be logged without introducing unrelated code in the bussiness layer
itself.