Implement example plugin for handling LoggingEvents
This plugin is just a very small example on how a plugin can be notified
about LoggingEvents raised in the core application and how to handle them.
The glue parts are the doWithSpring closure in the plugin description to
register a spring bean which implements the ApplicationListener. This
bean does not do anything useful except writing the received username
and method name to a logger.