Add ModelCreatedEvent which is raised after creating new Model
The ModelCreatedEvent is triggered by the ModelService when a user
successfully uploaded a new Model. It contains information about the
new Model and a reference to the File handle which contains the raw
Model data.

Interested parties can listen to the event and use it in any way
they want, though changing the data is not possible.