Properly implement the Publication information
The Publication has more normalized fields and a list of Authors. The Author and
Publication have related TransportCommandObjects as wrappers. A Publication has
a link and a link provider saying whether the link points to a PubMed resource,
is a DOI link or a "normal" link.

For the case of PubMed resources a PubMedService is added to retrieve information
from citexplore for the publication and set the appropriate fields. In case of
DOI or custom URL this is of course not possible.