Show Publication information in the Model overview table
Publication is added as a link to either the PubMed resource at citexplore
or as a link to the doi resource. The link title shows some basic information
about the publication, that is journal, issue, volume, pages, etc.
In a cluetip tooltip further information is presented nicely formatted in a table.
This includes the title, authors, affiliation, abstract. In fact all available
fields are presented.