Support Uploading new Models in Web Application
Initial support for an upload new Model view. Currently working if
the user selects a PubMed or an unpublished Publication.
DOI and custum URL and not yet implemented but UI is already there.

Proper error handling included. Validation on the server through AJAX.
Incorrect fields are highlighted and an error message is put to a site
global error field. The single error messages can be closed by the user.

For better AJAX handling the blockUI plugin is imported for blocking
the relevent UI parts while the upload is in progress. And the form
plugin is used for serializing the form. It is needed as we upload
a file which cannot be serialized with AJAX, therefore the plugin is
needed as it uses an iframe for uploading.

References:
blockUI: http://jquery.malsup.com/block/
form: http://jquery.malsup.com/form/