Implement proper error handling
An ErrorsController is added to handle server errors. It generates a digest from the exception
which can be used as a code to identify the exception. An error page is shown including the
code without including application relevant data (e.g. stacktrace).

In the AJAX case a JSON object is constructed and returned. The uploadModel AJAX callback is
modified to check for server errors and a handleError method is added which shows an error message
in the current view. With the (not yet included) new version 1.5 of jQuery this should be ported
to use the statusCode callback methods.