Loading the Model becomes more dynamic
JavaScript function added to load the list of models through AJAX and add it to the main view.
The actual list view data adjusted to only include the table but not boilerplate HTML.
Index page lost the grails stuff but now includes the div id=body element which is used
as target for the AJAX function.