Add Redirect Controller
This controller takes care of bringing the web application into the correct state,
so that we can have URI to all resources even if they are called by JavaScript.
Each controller needs to redirect to the home controller if a request is not an
AJAX request and the home controller will generate the correct JavaScript calls
depending on a passed identifier.