Handle access denied properly
In case of an AccessDeniedException we show a proper error message instead of the
generic unexpected error. The returned JSON contains information about the user's
logged-in status, so that the JavaScript can show the login dialog in case the user
is not authenticated. If the user is authenticated, he is missing the right and only
the error message needs to be shown.