Fetch Publication Tooltip data from AJAX
The Publication Information HTML is moved into an own template which is used in the summary view and
by a new controller action to just render the publication for a Model.

This allows us to have the code to build the Publication HTML only once, instead of twice. Next step is
to remove all the not needed Publication data before transmission to the Model view, to reduce overhead.