Add Perf4J integration
Perf4J [1] allows to get very concrete profiling data on the application execution.
It uses AspectJ to intercept method calls and log their performance. Added integration
to be used in the ModelService, JmsAdapterService and CoreAdapterService.

There exists a grails plugin for perf4j, but it does not work with our ModelService. It
uses groovy interception and this fails with the method ModelService.getAllModels.

What might be done: implement an own aspect to not need to add the @Profiled annotation
to all methods and maybe move Perf4J into an own plugin.

It should become runtime configurable or at least should be disabled in production mode.

[1]: http://perf4j.codehaus.org/