Log property is not available in afterPropertiesSet
The log property is injected into the services not before afterPropertiesSet is called.
Because of that application startup fails if the VcsService is not able to configure properly.
Of course the log messages are of value, therefore the change to use an init method and call it
from bootstrap as recommended in grails bug 3899.

http://jira.codehaus.org/browse/GRAILS-3899