Support for publication date specification
Year is a mandatory field, while month and date are optional.
Year is a dropdown with all years since 1980 till current year and month is a dropdown
with the short names for month. Day is just a dropdown with days from 1 to 31.
There is no logic yet to ensure that day is valid in the scope of month and year.