Support multiple author fields
The view is extended with an addAuthor button which will duplicate the author information table rows.
Additionally each new author segment has a remove author button which will remove the inserted rows again.
In the controller the authors are parsed before uploading the model, though proper validation and error
handling is not possible. The information is dynamic and cannot be encapsulated in the command object.