Show error icon next to each field with an error
The error icon has a title element with the actual error message.
The generic setErrorState JavaScript method takes care of adding the error
message as the title and showing/hiding the error icon.

In order to better position the element a css file is included which currently
just positions the input fields in a form.