Theme support for the web application
Jummp becomes fully themeable through jQuery-UI CSS framework. The themes have to be stored in
web-app/jquery-ui/<themename>/ and the CSS file needs to be named <themename>.css
The default theme is currently "smoothness". The default can be overwritten in the application.properties
through the "jummp.theme" property.

The user can overwrite the theme in each request with the "theme" parameter.
E.g. http://localhost:8080/jummp-web-application/?theme=themename

If the user selected or application configured themename is not valid the application falls back to the default smoothness.