User Registration
First part of user registration: service method which creates a new User.
At the current state it creates a user and either disables it or sets the
password to expired depending on a user registers himself or an administrator
creates the user. The password is only set if LDAP is not used as the
authentication backend.

A validation code is generated which is valid for one day. This validation
code can be used by the user to enable the account/set a password. Currently
the code for that is not yet written.

What needs to be done is to send out mails to the new registered user with
the validation code in place and the service method to validate the registration
needs to be written.