Retrieve configuration from Core in the Web Application
During boot strap of web application the core's configuration is loaded and added into the Web Application's
own configuration. This allows to have externalized configuration just loaded once (in the core) and use it
also in the web application.

E.g. we might not want to offer a change password functionality if the core uses LDAP as authentication
backend.