Config Option to disable changing of password
The new config option jummp.security.changePassword decides whether the UI should support changing of password.
The option defaults to true and is overwritten in case LDAP is used. We cannot (yet) change the password in LDAP
so we need to disable it.

If the option is disabled the view hides the change password fields and the action has a check to verify that changing
password is allowed.

Core is unchanged.