Reset Password Functionality
UserService is extended by two new methods: one to request a password reset.
It will send out a mail to the user (see new configuration options in config.groovy)
with a generated code to reset the password.

This code is used in conjunction with the username to allow to set a new password.

The approach should be secure. As the reset password code is sent to the email address
configured by the user, no attacker can just create a request code. An attacker needs
to have access to the mail account (out of our responsibility). If someone just gains
access to the reset code, he still needs to know the username.

What needs to be done for this part:
* unit test
* more specific Exceptions