Generic implementation for ajaxSubmit of forms
New method added which submits a form via AJAX and handles the response.
A passed in callback is invoked when the processing is successful.
All form submissions are changed to use this method and the callbacks are
moved into the view.js file (as it's view logic).

Currently it does not handle the file upload cases and when additional data
needs to be submitted to the form.