View for User Role management
The show user view in UserAdministrationController consists of two parts:
1. Normal user editing - reusing the /user/show/ view and action (views need a merge)
2. Role management. The Role management is shown as two lists: one with the roles the
user has, one with the roles the user does not have. Each role is rendered together
with a button to either add or remove the role. When clicking the button a request
to the server is sent and on success the role switches the list.