Better handling for non-ajax URL requests
All actions render a template and pass the AJAX request to be loaded as a model to the template.
This removes the string comparison which started to not scale anymore.
Direct transition to register is currently broken as it does not use the generic funtionality.
Using the normal link on login page works, of course.