Adding a Menu to the Web Application
The menu structure is defined in web-app/menu.xml and is security aware.
The xml structure of the menu is parsed in MenuService into a list of
MenuItems. Each MenuItem can contain another list of MenuItems to create
submenus.

Menus are cached for sets of user roles. This means the XML is only parsed
once for users with the same roles.

The menu is rendered into a div - ul - li - a stucture in the jummp taglib.
For better visualization the jdMenu jQuery plugin is used [1]. The menu is
rendered in the global document.ready event handler and reloaded after
login/logout - that is when the user's roles change. Therefore an additional
controler is added which just renders the menu through the taglib.

I will add an xsd to describe the xml structure.

[1]: http://jdsharp.us/jQuery/plugins/jdMenu/