Implement an AuthenticationHashService
The AuthenticationHashService has an internal Map with a hash as key and authentication as value.
It is possible to add an Authentication to the Map and a unique hash is generated and returned.
This can be used for the communication between core and the different application parts through
the IPC mechanism. By using the hash no actual user information is exchanged on the IPC bus.

If there is no Authentication for a given hash the service method throws an
AuthenticationHashNotFoundException which extends the AuthenticationException.