Adding a DBus interface to JUMMP core
DBus is the main IPC mechanism on modern Unix systems, but also supported on
Microsoft Windows and MacOS X. In opposite to JMS it is designed in a
synchronous manner and supports Exception states properly.

DBus is here meant to complement the JMS bindings and not to replace them. DBus
uses a unix socket for message passing, because of that it is a local bus - at least
on Unix systems.

The implementation so far exports to beans to DBus. One for accessing the Application
and one to access the UserService. Authentications are not passed around, but a token
is created and this is passed around to identify an Authentication. This makes the
DBus approach more secure than the JMS approach (which should be changed). DBus also
supports Peer-to-Peer communication (default is broadcast), but in general the bus
is trusted and many sensitive data is passed around in DBus in a normal Linux system.
Examples would be NetworkManager or KWallet. So we can safely assume that DBus is a
trusted environment and there is no need to implement the peer-to-peer communication.

The exported objects need to have a Java interface describing the methods exported
to DBus and the implementation needs to be a Java class as well. It seems like the
implementation uses introspection on the classes and this fails if the classes are
written in Groovy.

Exceptions thrown during method execution are translated into a DBusExecutionException.
If we want to have a more specific exception it needs to extend this exception.
Because of that this changeset includes also several new exception classes to map
the exception we are interested in the web application.

It is possible to serialize any object over DBus, but the object needs to implement
the DBusSerializable interface. Because of that new wrapper classes are introduced
to wrap the actual objects into DBus compatible objects.

The changeset does (not yet) include the libraries to run DBus. As it needs a
JNI component for the unix socket communication the dependencies have to be
compiled from source. Please download and follow the instructions on
http://dbus.freedesktop.org/doc/dbus-java/

The next step is to move the DBus part into an own plugin and ensure that DBus
is an optional dependency and add the web application part. If it is an optional
dependency it should be fine to include the dependency blobs (except the JNI library).