Fix ambiguous compilation error with Java and our Exceptions
Apparently Java fails to compile because constructors with one parameter all
are represented as parameter Object. In case of Jummp Exception removed the
constructor with just a Throwable (no inhereting class uses it), in case of
UserManagementException switch from Long to long and fix a bug there at the
same time.