Adding interface and delgating service for ModelService
Extracted an interface IModelService from ModelService. As the interface should
be in core-api it is changed to handle CommandObjects instead of the DOM classes.
An additional service is added wich implements the interface and delegates all
methods calls to the ModelService and handles the translation to and fro
CommandObjects.

This allowed to change the ModelJmsAdapterService to no longer translate to
the CommandObjects.