Adding an own Authentication Implementation with the AuthenticationHash
New interface JummpAuthentication extends Authentication and adds a getter
for the Authentication Hash. The DBusAuthentication is changed to implement
this interface and a concrete implementation is also added to be used by
JMS and passed around by JMS.

Next step: change JMS to use the authentication hash instead of Authentication.