JMS uses the Authentication Hash to identify an Authentication
This is an important change in the Authentication process. JMS does no longer
pass the complete Authentication Object to the core, but only an identifying hash,
just as DBus does. This means that it is no longer possible to inject Authentications
into core and makes the process more secure.

The only insecure remaining part is that the initial authentication is unencrypted as well
as the Authentication Hash. This means eavesdropping the connection is still vulnerable to
security.