Use new statusCode in ajax requests
Status code allows us to have a function executed per different status.
Added 403, 404 and 500 handlers. The server does not need to response
with a JSON object but just text with the info required. E.g. in case
of 404 the not found resource, in 500 the error code and in 403 whether
the user is logged in or not (allowing to show the authentication dialog).