Retrieve Model Files through DBus
Instead of byte array in DBus we move only the path to the file around as we know
that we are on a local system and our two components executed with the same user
priveledges.

Exceptions are not yet wrapped and it could make sense to change ModelService to
pass File handles around instead of the byte array.