RemoteModelAdapter takes ids as parameters and not ModelTransportCommands
This slightly improves the way how to interact with the RemoteModelService as
the controllers only need to pass the ids to the methods and not wrap it inside
a ModelTransportCommand. In case of DBus the additional unwrapping is no longer
required and for JMS it would make sense to change it as well (less overhead).