Using a factory class to create the GitManager
In fact the GitService was used like a factory class, so let's make a factory out of it.
The new GitManagerFactory still uses the contract of VCS and can be used in the same way
as the service before. But it is no longer a Grails service and configured in the plugin.
This allows to move the complete configuration into the plugin.

The next step is to similar change the SVN plugin and then to dependency inject the VcsManager
into the VcsService.