Adding a most recently used cache for SBMLDocuments
The cache is a simple Map using a RevisionTransportCommand as the
key and an SBMLDocument as the value. The cache is intended for
keeping recently used SBML Models parsed instead of re-parsing
each time it is required.

The implementation is thread safe.

Currently not all methods are properly tested yet.