Begin SBML parsing fun
SbmlService provides first methods to parse model information like
level, version and metaId on model level. The service uses the new
cache to not re-parse the file each time a model is accessed.

A new integration test is added to core to test these new methods.
The test cannot be in the plugin as we need access to core services
as we need to access the underlying VCS to retrieve model files if
not in the cache (this is of course protected by using the ModelService).