Export SbmlService to DBus
What we currently have for extracting information from SBML is exported to DBus.
The dbus interface is added in the normal DBus package and an interface for
accessing the SBMLService is added to core-api. This slightly violates the
plugin approach. Given that sbml support is in a plugin the interface and
dbus interface should not be in the core parts. But adding an own plugin
just for one interface seems overhead as well.

An idea for improvement is to not register the DBus interface if the
sbmlService is not available.