Build up a runtime database about MIRIAM resources
During bootstrap the new MiriamService is invoked and the resource
description is fetched, parsed and stored in Domain objects. As the
web-application does not have a direct database, everything is in
memory. Given that the dataset is rather small, there should not
be a problem with that.

For storing two Domain classes are used: MiriamDatatype to describe
one datatype and MiriamResource to describe one resource where the
datatype can be resolved.