Render Model annotations
In order to render annotations new tags are added to the JummpTaglib.
The Taglib uses the MiriamService to resolve the correct datatype and
render the name and link to the resource. Currently the first, not
obsoleted resource is used. The existing Biomodels DB has a static map
for finding the preferred resource, this should be somehow added (but
in a more flexible way than writing into the source code).

A way to resolve the name of a given URI is also required.