Resolve names from EBI webservices for MIRIAM URNs
Added a way to resolve the names for a given URN. Unfortunately we need to
hardcode as we have to download and parse XML. Each term is only downloaded
once and added to a hash of resolved terms.

Currently added support for Taxonomy and Gene Ontology - more to come.

In case the EBI wants to continue to directly connect to databases, the code
to resolve could be moved into a plugin and replaced by another plugin which
connects to the databases.