Adding a view to render SBML parameters
The view is highly inspired from the existing Biomodels DB page (surprise, surprise), but
with some variations. The Notes and Annotations are shown inside a tooltip instead of
an expandable element.
Currently only global parameters are rendered, annotations (except the SBO term) are still
missing and the link to assignment rules is not yet added (given that we do not yet render
assignment rules).