Retrieve listOfParameters of reactions
Additional to the global paramaters, there are local parameters on reactions. This addition
allows to retrieve all local parameters of the SBML file. The paramters are grouped by their
reaction and are wrapped in an additional list for the reaction. Each reaction is described
by a map with id, name and the parameters.

In comparison to the global parameters a local parameter is always a constant. The parameterToMap
method of SbmlService is adjusted to take care of this detail.