Render local parameters in the parameter view
The view logic is generalized and moved into a taglib so that we can
render both the global parameters as well as the local parameters grouped
by their reactions.

The code in the taglib is currently a little bit ugly as it generates HTML
code from within groovy code. I'll look into either using a HTML template or
a builder.

For the local parameter the tooltip is not yet working correctly, probably
needs fixing in core.