Enables check of users' authentication hashes through web application.
When a user's authentication is removed (i.e. because inactive period),
the web application does not know about the removal and tries to act
as before. This results in exceptions, when functionalities are tried
to access which are no more accessible.
The added functionalities provide an API to enable checking, if a
user's authentication still is valid, from a remote plugin.

Methods are added to the JMS and D-Bus systems.