Begin rendering the Math tab
We retrieve the list of Reactions and render them in the view.
Currently only the name and the simple visualization is shown.
Rendering of Math, Annotation and notes is still missing and
will be done in a tooltip similar to the parameters.

Also only reactions are rendered. This means no code yet for
rules, events and functions.

This first extension shows that code written for the parameters
needs to be made more generic.