Custom filter clearing the context.
This custom filter cleares a user's context,
if authentications are removed by the core.
It is registered in the filter chain after
SECURITY_CONTEXT_FILTER.
The filter ensures, that users' authentications
which are due to inactive period removed,
do not throw exceptions in the
web application and display the logged off
users the login dialog.