Convert content MathML to presentation MathML
Using the xls template provided by JEuclid for the conversion (Apache License).
Modern browsers are able to render presentation MathML without adjustments, but
content MathML is not rendered. Because of that a new tag is added to the jummp
TagLib to convert the content MathML using xslt to presentation MathML and just
writing it to the generated document.

This works fine with Firefox (4) which does have native MathML support, but
does not work in WebKit browsers (tested rekonq and Chromium). For such browsers
mathJax [1] might be a solution.

[1]: http://www.mathjax.org/