Retrieve all rules from the SBML document
Rules are rather tricky as the rules come in three different forms as
assignment rules, rate rules and algebraic rules. The access methods
between the various forms differ and they are alltogether different
to the other SBML elements, e.g. id and name cannot be extracted.

The returned map tries to get one common way to represent the information
by adding some type information and passing variable information in a
standardized way to the map.