Render rules in the web interface
As rules come in different forms we have one tag and gsp for each kind
of rule. Unfortunately I have not yet found a model with an algebraic
rule, so this is untested.

The rule representation differs slightly to our other elements by not
adding tooltips to it. BiomodelsDB provides the collapsed row, but
only with redundant information - that is the MathML which is already
shown in the rule field. Because of that a tooltip seems superfluous.