Adding AST Transformation to generate the code of the Remote*Services
An AST Transformation allows to inject new code during the compile phase.
See http://groovy.codehaus.org/Compile-time+Metaprogramming+-+AST+Transformations
for further information.

The Remote*Services have very similar code by just implementing an interface and
delegating all methods of the interface to a property of the same interface. This
can be simplified by the code generation. The methods are now generated during the
compile time.

As the transformation has to be compiled before the sources which need to be modified
are compiled, we need to have the AST Transformation in an own plugin which is compiled
through the build dependency before the code to be transformed.